% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment.R
\name{catchment_area}
\alias{catchment_area}
\alias{.catchment_area}
\alias{.catchment}
\alias{.do_catchment}
\title{Catchment worker functions
Worker functions for catchments and catchment areas on single pixels}
\usage{
.catchment_area(y, out_name, drain_name)

.catchment(y, out_name, drain_name)

.do_catchment(y, out_name, drain_name)
}
\arguments{
\item{y}{Coordinates, where to compute the catchment}

\item{out_name}{Name of the output raster in grass}

\item{drain_name}{Name of the (already present) drainage direction raster}
}
\description{
Catchment worker functions
Worker functions for catchments and catchment areas on single pixels
}
\details{
\itemize{
\item \code{.do_catchment} builds the catchment raster in GRASS
\item \code{.catchment_area} runs \code{.do_catchment} and then comptues and returns the area as a number
\item \code{.catchment} runs \code{.do_catchment} and then extracts and returns the raster from GRASS
}
}
\keyword{internal}
