% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{delineate}
\alias{delineate}
\title{Delineate streams from an elevation model}
\usage{
delineate(
  dem,
  threshold = 1e+06,
  pretty = FALSE,
  file,
  outlet = NA,
  reach_len,
  ...
)
}
\arguments{
\item{dem}{A \link[raster:raster]{raster::raster}; a digital elevation model.}

\item{threshold}{The minimum size of a delineated catchment; units are in dem_units^2 (e.g., m^2); default 10 km^2.}

\item{pretty}{Should flat areas be made prettier? Corresponds to r.watershed -b flag}

\item{file}{The file name of the raster to be returned, see \code{details}.}

\item{outlet}{The location of the outlet of the target stream, see \code{details}.}

\item{reach_len}{Optional; if provided, reaches will be resized to meet this target length}

\item{...}{If \code{reach_len} is specified, additional parameters to be passed to \code{\link[=resize_reaches]{resize_reaches()}}}
}
\value{
A \code{\link[raster:stack]{raster::stack()}}, containing the drainage map, the flow accumulation map, and the delineated streams.
}
\description{
Delineate streams from an elevation model
}
\details{
This is a wrapper for \href{https://grass.osgeo.org/grass76/manuals/r.watershed.html}{r.watershed}.

The threshold parameter controls the level of detail in the delineated streams. Higher thresholds result in faster
computation and fewer streams in the output. For finer control, see \code{\link[=extract_stream]{extract_stream()}}.

IIf \code{outlet} is \code{NA} (the default), then the largest stream in the area will be set as the outlet,
and only streams in that watershed will be used.
If a smaller stream is the focus, then outlet can be a pair of x-y coordinates which will determine the farthest
downstream point to use.

Streams can be converted to a vector file, see \code{\link[=stream_vector]{stream_vector()}}.

It is recommended to specify the \code{file} parameter (including the extension to specify
file format; e.g., .tif, .grd). If not specified, a temp file will be created and will be
lost at the end of the R session.
}
\examples{
\donttest{
    data(kamp_dem)
    x = delineate(kamp_dem)
}
}
