% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{.add_raster}
\alias{.add_raster}
\title{Add a raster to grass}
\usage{
.add_raster(x, name, flags, overwrite = TRUE)
}
\arguments{
\item{x}{A \link[terra:SpatRaster-class]{terra::SpatRaster}}

\item{name}{The name of the layer in grass}

\item{flags}{Any flags to pass to \link[rgrass:readRAST]{rgrass::write_RAST}}

\item{overwrite}{Should the file be overwritten if it exists}
}
\description{
Add a raster to grass
}
\keyword{internal}
