% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{.check_topology}
\alias{.check_topology}
\title{Verify that the topology is working as expected}
\usage{
.check_topology(Tp, warn = FALSE)
}
\arguments{
\item{Tp}{A topology matrix}

\item{warn}{if TRUE, warnings are raised instead of errors}
}
\value{
NULL, raises errors/warnings when invalid topologies detected
}
\description{
Verify that the topology is working as expected
}
\keyword{internal}
