% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{.clean_grass}
\alias{.clean_grass}
\title{Clean up grass files}
\usage{
.clean_grass(raster, vector)
}
\arguments{
\item{raster}{Raster layers to remove, if missing removes all, if NA removes none}

\item{vector}{Vector layers to remove, if missing removes all, if NA removes none}
}
\description{
Clean up grass files
}
\details{
The default mode is to clean everything, removing all layers from the grass session
}
\keyword{internal}
