% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.r
\name{.do_intersect}
\alias{.do_intersect}
\title{Helper function to perform the watershed intersection on a single reach}
\usage{
.do_intersect(r, x, y, width, riv_buff, rid, dname)
}
\arguments{
\item{r}{A single reach, an sf linestring}

\item{x}{A single point, the outlet of r}

\item{y}{A raster stack to intersect}

\item{width}{Buffer width}

\item{riv_buff}{A buffer layer for the entire river}

\item{rid}{The id number of the reach}

\item{dname}{The name of the drainage direction raster in grass}
}
\description{
Helper function to perform the watershed intersection on a single reach
}
