% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{.fix_drainage}
\alias{.fix_drainage}
\title{Fix drainage direction for a single pixel}
\usage{
.fix_drainage(id, drain, conn)
}
\arguments{
\item{id}{ID of the problematic pixel}

\item{connMat}{preliminary connectivity matrix}

\item{drainMat}{Drainage direction matrix}
}
\value{
corrected row from the connectivity matrix
}
\description{
Fix drainage direction for a single pixel
}
\keyword{internal}
