% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{.fix_topology}
\alias{.fix_topology}
\title{Check and fix problems with drainage direction}
\usage{
.fix_topology(conn, drain, prev_probs = NA)
}
\arguments{
\item{conn}{A preliminary topology matrix}

\item{drain}{Drainage direction matrix}

\item{prev_probs}{Previous number of problems, to allow stopping if no improvement on
subsequent calls}
}
\value{
A corrected topology matrix
}
\description{
Check and fix problems with drainage direction
}
\details{
In some cases, drainage direction rasters don't agree with flow accumulation, resulting
in a delineated stream that doesn't have the right drainage direction. This function
attempts to detect and fix this in the adjacency matrix and the drainage layer.\preformatted{	Not used in the current version; can be reactivated if there are problems
}
}
\keyword{internal}
