% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{.flowto}
\alias{.flowto}
\title{Compute which pixels flow into which other pixels}
\usage{
.flowto(mat, xmax, ymax)
}
\arguments{
\item{mat}{A matrix with 4 named columns; 'x', 'y', 'drainage', and 'id'}
}
\value{
A matrix of IDs, the first column the upstream pixel, the second column downstream pixel
}
\description{
Compute which pixels flow into which other pixels
}
\keyword{internal}
