% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{.preferred_grass_version}
\alias{.preferred_grass_version}
\title{Choose a preferred version of grass from a list of options}
\usage{
.preferred_grass_version(x)
}
\arguments{
\item{x}{A vector of grass home directories}
}
\value{
The most preferred from x
}
\description{
Choose a preferred version of grass from a list of options
}
\keyword{internal}
