% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.r
\name{.process_area}
\alias{.process_area}
\title{Take whatever format of areal layer, process into a raster for ws_intersect}
\usage{
.process_area(x, ...)
}
\arguments{
\item{x}{The areal layer to process}

\item{id}{The id column in x, needed if x is a vector GIS layer}

\item{ras}{The raster template to use to create a raster from x}

\item{return}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}}, with each layer classified, optionally preserving names
from vector layers}
}
\description{
Take whatever format of areal layer, process into a raster for ws_intersect
}
\keyword{internal}
