% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{.raster_to_spdf}
\alias{.raster_to_spdf}
\title{Produce a \link[sp:SpatialGridDataFrame]{sp::SpatialPixelsDataFrame} from a \link[raster:raster]{raster::raster}}
\usage{
.raster_to_spdf(x, complete.cases = FALSE)
}
\arguments{
\item{x}{A \link[raster:raster]{raster::raster} object}

\item{complete.cases}{Boolean, if TRUE only complete (non-na) rows are returned}
}
\value{
A \link[sp:SpatialGridDataFrame]{sp::SpatialPixelsDataFrame}
}
\description{
Produce a \link[sp:SpatialGridDataFrame]{sp::SpatialPixelsDataFrame} from a \link[raster:raster]{raster::raster}
}
\keyword{internal}
