% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{.read_rasters}
\alias{.read_rasters}
\title{Read and format rasters from a grass session}
\usage{
.read_rasters(layers, file)
}
\arguments{
\item{layers}{A vector of names of rasters to read}

\item{file}{The file name to save the raster}
}
\description{
Read and format rasters from a grass session
}
\keyword{internal}
