% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{.resize_r}
\alias{.resize_r}
\title{Split a single reach}
\usage{
.resize_r(ids, Tp, trim, len, min_len)
}
\arguments{
\item{ids}{A vector of pixel ids in the reach, sorted from upstream to downstream}

\item{Tp}{A pixel topology}

\item{trim}{Logical, should we trim small headwater reaches?}

\item{len}{Target reach length}

\item{min_len}{Minimum acceptable reach length, only used if \code{trim} is \code{TRUE}}
}
\value{
A list of vectors, each vector is a single new reach.
}
\description{
Split a single reach
}
\keyword{internal}
