% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{.start_grass}
\alias{.start_grass}
\title{Start grass in a simple way}
\usage{
.start_grass(
  layer,
  layer_name,
  gisBase,
  home = tempdir(),
  gisDbase = home,
  location = "watershed",
  mapset = "PERMANENT"
)
}
\arguments{
\item{layer}{A \link[raster:raster]{raster::raster} object to write to the grass session}

\item{layer_name}{Name of the layer in the grass environment}

\item{gisBase}{character; the location of the GRASS installation}

\item{home}{Location to write GRASS settings}

\item{gisDbase}{Location to write GRASS GIS datasets}

\item{location}{Grass location name}

\item{mapset}{Grass mapset name}
}
\description{
Start grass in a simple way
}
\keyword{internal}
