% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.r
\name{fill_dem}
\alias{fill_dem}
\title{Produce a DEM with basins filled in}
\usage{
fill_dem(dem, file = NA)
}
\arguments{
\item{dem}{A \link[raster:raster]{raster::raster}; a digital elevation model.}

\item{file}{The file name of the raster to be returned, see \code{details}.}
}
\value{
A filled DEM raster
}
\description{
Produce a DEM with basins filled in
}
\details{
This is a wrapper for \href{https://grass.osgeo.org/grass76/manuals/r.fill.dir.html}{r.fill.dir}

It is recommended to specify the \code{file} parameter (including the extension to specify
file format; e.g., .tif, .grd). If not specified, a temp file will be created and will be
lost at the end of the R session.
}
\examples{
\donttest {
    data(kamp_dem)
    kamp_fill = fill_dem(kamp_dem)
}
}
