% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaa.R
\name{get_nwaa_models}
\alias{get_nwaa_models}
\title{Fetch Available USGS NWAA Models}
\usage{
get_nwaa_models()
}
\value{
A \code{data.frame} describing available NWAA models and descriptions.
}
\description{
Retrieves metadata for available scientific models in the USGS NWAA Data Companion API.
}
\examples{
\dontrun{
models <- get_nwaa_models()
}
}
