% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaa.R
\name{get_nwaa_variables}
\alias{get_nwaa_variables}
\title{Fetch Available USGS NWAA Variables}
\usage{
get_nwaa_variables()
}
\value{
A \code{data.frame} of available NWAA variables with identifiers and descriptions.
}
\description{
Retrieves metadata for available environmental and hydrological assessment variables
in the USGS NWAA Data Companion API.
}
\examples{
\dontrun{
vars <- get_nwaa_variables()
}
}
