% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{hydraulic_geometry}
\alias{hydraulic_geometry}
\title{Compute discharge and hyrdaulic geometry}
\usage{
hydraulic_geometry(Ac, Qc, Ap)
}
\arguments{
\item{Ac}{Vector of catchment area for calibration points in $m^2$}

\item{Qc}{Vector of discharge for calibration points in $m^3/s$}

\item{Ap}{Vector or raster of catchment area for prediction}
}
\value{
A data.frame with the following variables:
* CA; the input catchment area
* Q; discharge
* v; water velocity
* z; water depth
* w; width
* Ax; stream cross-sectional area
}
\description{
Compute discharge and hyrdaulic geometry
}
\details{
Computes discharge from catchment area as \eqn{\log Q = \log b + m \log A}.
If a single point in \code{calib} is included, the relationship will be re-parameterised by
adjusting the intercept parameter \code{b} so that the calibration point falls on the line
(while keeping the slope the same).\preformatted{	With multiple points, a bayesian regression model is fit with rstanarm. The model uses
	Burgers et al parameters as an informative prior.

	The default parameters used by this function come from Burgers et al (2014).
	For these parameters, catchment area units are expected to be in
	\eqn{m^2}, and discharge will be computed in \eqn{m^3 s^{-1}}.
}
}
\examples{
\donttest {
    library(sf)
    data(kamp_q)
    data(kamp_dem)

    kamp = delineate(kamp_dem)
    kamp_Tp = pixel_topology(kamp)
    CA = catchment(kamp, type = 'reach', Tp = kamp_Tp)

    ## need to transform coordinate system for kamp_q to match
    ## need to snap to stream
    ## need to make sure this works
    ## should probably update kamp_q projection to match kamp_dem

    kamp_q$ca = catchment(kamp, type = 'points', y = st_geometry(kamp_q), Tp = kamp_Tp)
    x = hydraulic_geometry(kamp_q$ca, kamp_q$discharge, CA)
}
}
\references{
Burgers HE et al. 2014. Size relationships of water discharge in rivers: scaling
of discharge with catchment area, main-stem lengthand precipitation.
\emph{Hydrological Processes}. \strong{28}:5769-5775.

Raymond, PA et al. 2012. Scaling the gas transfer velocity and hydraulic
geometry in streams and small rivers. \emph{Limnology and Oceanography: Fluids and
Environments}. \strong{2}:41-53.
}
