% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\docType{data}
\name{kamp_dem}
\alias{kamp_dem}
\title{Digital elevation model for the Kamp River catchment in Austria}
\format{
A \code{\link[raster:raster]{raster::raster()}}
\describe{
\item{value}{Elevation of each pixel, in m}
}
}
\usage{
kamp_dem
}
\description{
Digital elevation model for the Kamp River catchment in Austria
}
\keyword{datasets}
