% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\docType{data}
\name{kamp_q}
\alias{kamp_q}
\title{Measured discharge in the Kamp river}
\format{
A simple features (\code{sf}) point dataset
\describe{
\item{discharge}{Discharge in m^3/s}
}
}
\usage{
kamp_q
}
\description{
Measured discharge in the Kamp river
}
\keyword{datasets}
