% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{load_watershed_data}
\alias{load_watershed_data}
\title{Load in and parse Watershed input file}
\usage{
load_watershed_data(
  input_file,
  number_of_dimensions,
  pvalue_fraction,
  pvalue_threshold
)
}
\arguments{
\item{input_file}{String corresponding to the file name of the Watershed input file.
Either a file path or a URL.}

\item{number_of_dimensions}{Integer representing the number of outlier types. 
Sometimes referred to as \code{E} in our documentation.}

\item{pvalue_fraction}{Float. Fraction of outliers (based on rank) that are 
considered an outlier for N2 pair analysis (this is done so each outlier type/signal 
has approximately the same distribution of positive outlier examples).}

\item{pvalue_threshold}{Float. Absolute p-value threshold used to create 
binary outliers used for Genomic Annotation Model.}
}
\value{
A named list
}
\description{
Load in and parse Watershed input file
}
\examples{
input = paste0("https://raw.githubusercontent.com/BennyStrobes/Watershed/",
     "master/example_data/watershed_example_data.txt")
data <- load_watershed_data(input, 3, 0.1, 0.1)

}
