% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{pixel_topology}
\alias{pixel_topology}
\title{Buid a topology for each pixel in a delineated stream}
\usage{
pixel_topology(x, drainage, stream)
}
\arguments{
\item{x}{A \link[raster:stack]{raster::stack}, such as one created by \code{\link[=delineate]{delineate()}}, or specify layers separately with \code{drain} and \code{stream}.}

\item{drainage}{Optional, ignored if \code{x} is provided; a drainage direction raster}

\item{stream}{Optional, ignored if \code{x} is provided; a delineated stream raster, all non-stream cells must be \code{NA}}
}
\value{
A \link[Matrix:sparseMatrix]{Matrix::sparseMatrix} giving the pixel topology
}
\description{
Buid a topology for each pixel in a delineated stream
}
