% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.r
\name{q_calib}
\alias{q_calib}
\alias{.q_calib}
\alias{.q_prior}
\alias{.q_predict}
\alias{.q_lpost}
\title{Calibrate discharge
Runs a bayesian linear regression to compute discharge from catchment area}
\usage{
.q_calib(a, q)

.q_prior()

.q_predict(par, a)

.q_lpost(par, a, q)
}
\arguments{
\item{x}{A \link{raster} stream map, such as one created by \code{\link[=delineate]{delineate()}}.}
}
\value{
An \code{sf} stream layer
}
\description{
Calibrate discharge
Runs a bayesian linear regression to compute discharge from catchment area
}
\keyword{internal}
