% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.r
\name{read_rasters}
\alias{read_rasters}
\alias{.read_rasters}
\alias{.read_vector}
\title{Read files from grass
Read and format raster and vector layers from a grass session}
\usage{
.read_rasters(layers, file)

.read_vector(x)
}
\arguments{
\item{layers}{A vector of names of rasters to read}

\item{file}{The file name to save the raster}

\item{x}{A single vector layer name to read}
}
\description{
Read files from grass
Read and format raster and vector layers from a grass session
}
\keyword{internal}
