% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem.r
\name{river_slope}
\alias{river_slope}
\title{Compute the slope along the river channel}
\usage{
river_slope(x, dem)
}
\arguments{
\item{x}{A river vector layer in sf format, e.g., output from \code{\link[=vectorise_stream]{vectorise_stream()}}}

\item{dem}{A digital elevation model, preferably a \link[terra:SpatRaster-class]{terra::SpatRaster}}
}
\value{
Vector of slopes, one per reach, map units (e.g., meters of drop per meter of stream)
}
\description{
Computes the change in elevation from one reach to the next.
}
\examples{
\donttest{
    data(kamp_dem)
    x = delineate(kamp_dem)
    vect = vectorise_stream(x$stream, pixel_topology(x))
   vect$slope = river_slope(vect, kamp_dem)
}
}
