% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strahler.r
\name{strahler}
\alias{strahler}
\title{Compute Strahler stream order}
\usage{
strahler(Tx)
}
\arguments{
\item{Tx}{a (pixel or reach) topology}
}
\value{
A vector of stream orders
}
\description{
Compute Strahler stream order
}
