% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{stream_coordinates}
\alias{stream_coordinates}
\title{Return coordinates of just the wet pixels from a stream stack}
\usage{
stream_coordinates(x)
}
\arguments{
\item{x}{A stream stack}
}
\value{
A matrix of coordinates
}
\description{
Return coordinates of just the wet pixels from a stream stack
}
