% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{upstream}
\alias{upstream}
\alias{.upstream}
\alias{.downstream}
\alias{.headwater}
\alias{.outlet}
\alias{.is_headwater}
\alias{.is_outlet}
\alias{.upstream_r}
\title{Simple topology analysis
Simple functions for extracting information from river topologies}
\usage{
.upstream(Tx, j)

.downstream(Tx, i)

.headwater(Tx)

.outlet(Tx)

.is_headwater(Tx)

.is_outlet(Tx)

.upstream_r(i, stream, Tx)
}
\arguments{
\item{Tx}{A (pixel or reach) topology}

\item{i, j}{Focal node}

\item{stream}{A stream raster (as produced by \code{\link[=delineate]{delineate()}}) with reachIDs as the values}
}
\value{
The id (corresponding to dims of Tx) of the desired node(s)

For \code{.upstream_r}, A named2-column matrix, 'to' is the downstream reach (i.e., i) and
'from' contains ids of the reach(es) upstream of i; or NULL if there are none
}
\description{
Simple topology analysis
Simple functions for extracting information from river topologies
}
\details{
upstream/downstream take a single node as input, and return the up- or downstream node.
Outlets/headwaters analyzes the entire toplogy and finds nodes that have no downstream or upstream neighbours.
}
\keyword{internal}
