% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{vectorise_stream}
\alias{vectorise_stream}
\title{Vectorize a stream layer
Produces a vector layer (in \code{sf} format) from a raster stream map as created by \code{\link[=delineate]{delineate()}}.}
\usage{
vectorise_stream(x, Tp)
}
\arguments{
\item{x}{A \link{raster} stream map, such as one created by \code{\link[=delineate]{delineate()}}.}

\item{Tp}{A pixel topology}
}
\value{
An \code{sf} stream layer
}
\description{
Vectorize a stream layer
Produces a vector layer (in \code{sf} format) from a raster stream map as created by \code{\link[=delineate]{delineate()}}.
}
\examples{
\donttest{
    data(kamp_dem)
    x = delineate(kamp_dem)
    vect = vectorise_stream(x$stream, pixel_topology(x))
}
}
