% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{vectorise_stream_old}
\alias{vectorise_stream_old}
\title{Vectorize a stream layer
Deprecated, use vectorise_stream()..}
\usage{
vectorise_stream_old(x)
}
\value{
An \code{sf} stream layer
}
\description{
Vectorize a stream layer
Deprecated, use vectorise_stream()..
}
\keyword{internal}
