% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.r
\name{w_intersect}
\alias{w_intersect}
\title{Intersect a river feature with a polygon feature}
\usage{
w_intersect(
  riv,
  areas,
  area_id = NULL,
  drainage,
  pts,
  buff = 50,
  rid = "reach_id"
)
}
\arguments{
\item{riv}{An sf LINESTRING layer, the river channel}

\item{areas}{Areal features to summarize; see 'details.'}

\item{area_id}{Optional character vector giving field/column names that should be summarised;
see 'details.'}

\item{drainage}{Drainage direction raster for calculating catchment}

\item{pts}{An \code{sf} point layer, one point per reach; the locations to compute the catchment of
each reach. If missing, the last point in each reach in \code{riv} is used.}

\item{buff}{Buffer width}

\item{rid}{The column name in 'riv' containing reach ID numbers}
}
\value{
A data.table summarising the polygons in each layer in areas along the river provided in x.
}
\description{
w_intersect handles the use case where we want to summarize areal features,
such as land use, soils, etc, across reaches of a river, while taking into account drainage area,
flow, etc.
}
\details{
There are three output types:
\itemize{
\item \code{riparian}: The layers in \code{areas} are summarized within a riparian buffer around each reach
\item \code{riparian_upstream}: The layers in \code{areas} are summarized within a riparian buffer for each
focal reach and all upstream reaches
\item \code{catchment}: The layers in \code{areas} are summarized within the entire catchment for each reach
}

\code{areas} must be in one of the following formats: \code{terra::SpatRaster}, \code{raster::raster},
\code{sp::SpatialPolygons}, or \code{sf::sf}. Multiple features are allowed, in which case provide a
raster stack/multi band raster, or a list of polygon features.

If polygon features are specified in \code{areas}, you must also specify which attributes should be
summarised in \code{area_id}. Currently a single attribute per feature is supported, and it must
be specified by name, not column index. Partial matching with regular expressions is allowed, but
each item in \code{area_id} must identify a single column in \code{areas}. Ignored if \code{areas} is a
raster.
}
