% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watershed-package.R
\docType{package}
\name{Watershed-package}
\alias{Watershed}
\alias{Watershed-package}
\title{Watershed: Prioritization of functional rare genetic variants by integrating genomic annotations and RNA-seq}
\description{
Watershed is an unsupervised probabilistic framework that integrates genomic annotations and RNA-seq outlier calls to identify the probability a rare variant has a functional effect on a particular RNA-seq outlier phenotype (examples of outlier phenotypes can be, but are not limited to total expression, splicing, or ASE). Watershed extends our previous model RIVER (which can also be run via this package) by incoorperating information from multiple outlier phenoytpes into one model, where predictions for functional effects in one outlier phenotype are informed by observed outlier calls in another phenotype.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicolerg/Watershed}
  \item \url{https://github.com/BennyStrobes/Watershed}
  \item \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7646251/}
}

}
\author{
\strong{Maintainer}: Nicole Gay \email{nicole.r.gay@gmail.com} [contributor]

Authors:
\itemize{
  \item Ben Strober \email{bstrober3@gmail.com}
}

}
\keyword{internal}
