% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\docType{package}
\name{watershed_package}
\alias{watershed_package}
\title{watershed: Tools for Watershed Delineation}
\description{
This is a simplified set of tools for deriving a river network from digital elevation data.
The package relies on GRASS GIS (v. 7.4 or 7.6) to do the heavy lifting. An installation of
GRASS and the rgrass7 package is required for watershed to function. Note that this is meant
to be a simplified workflow for watershed delineation using only R code; for more options
users can use rgrass7 to access GRASS GIS functions directly.
}
\section{Key functions}{

\itemize{
\item \code{\link[=delineate]{delineate()}} Produce a stream map from a digital elevation model
\item \code{\link[=pixel_topology]{pixel_topology()}} Construct river network topologies
\item \code{\link[=reach_topology]{reach_topology()}} Construct river network topologies
\item \code{\link[=catchment]{catchment()}} Compute catchment areas
}
}

\section{Datasets}{

\itemize{
\item \link{kamp_dem} An example digital elevation model for the Kamp river in Austria
\item \link{kamp_q} Measured discharge in the Kamp river
}
}

