% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.r
\name{wsi_summary}
\alias{wsi_summary}
\alias{.wsi_summary}
\alias{.wsi_summary_single}
\title{Summarise multiple raster layers by area}
\usage{
.wsi_summary(ras, ...)

.wsi_summary_single(ras)
}
\arguments{
\item{ras}{A raster stack}

\item{...}{Additional (single-valued) attributes to add to the output table}
}
\value{
A data.table with the following columns:
\itemize{
\item method: The method used, either 'riparian', 'riparian_upstream', or 'catchment'
\item layer: The layer of origin for each row
\item value: The numeric value from the raster layer of each row
\item area: The area occupied by the category
\item proportion: The proportion of area (within the method and layer)
}
}
\description{
Summarise multiple raster layers by area
}
\keyword{internal}
