% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletApp.R
\name{leafletOutput}
\alias{leafletOutput}
\alias{leafletServer}
\alias{leafletApp}
\title{Interactive Leaflet Mapping UI (Output)}
\usage{
leafletOutput(id)

leafletServer(id, huc_level = 8, max_hucs = 6)

leafletApp()
}
\arguments{
\item{id}{Module ID}

\item{huc_level}{Target USGS HUC digit level (2-12, default: 8, can be numeric or reactive).}

\item{max_hucs}{Maximum target number of HUC regions when searching drawn polygon extent (default: 6, can be a numeric or reactive).}
}
\value{
A list of reactive objects: `huc` (reactiveVal holding discovered `sf` HUC polygon(s)),
  `status` (reactiveVal holding HTML status message), `click` (reactive holding map click details),
  and `drawn_polygon` (reactiveVal holding user drawn rubberband polygon sf).
}
\description{
Interactive Leaflet Mapping UI (Output)

Interactive Leaflet Mapping Server Logic
Server logic for interactive Leaflet discovery. Returns a list of reactives
(`huc`, `status`, `click`, `drawn_polygon`) enabling Shiny module composition.

Run the Leaflet Mapping Integration App Tracker
}
