% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.r
\name{resize_reaches}
\alias{resize_reaches}
\title{Resize reaches from a delineated stream}
\usage{
resize_reaches(x, Tp, len, min_len = 0.5 * len, trim = TRUE)
}
\arguments{
\item{x}{A \link{raster} stream map, such as one created by \code{\link[=delineate]{delineate()}}.}

\item{Tp}{A pixel topology}

\item{len}{The target length of resized reaches (in map units)}

\item{min_len}{How small is the smallest acceptable reach?}

\item{trim}{Logical; remove headwater reaches smaller than min_length?}
}
\value{
A raster similar to \code{x} with reaches resized
}
\description{
Resize reaches from a delineated stream
}
\examples{
\donttest{
    data(kamp_dem)
    x = delineate(kamp_dem)
    Tp = pixel_topology(x)
    x[['stream']] = resize_reaches(x[['stream']], Tp, 1000, 600)
}
}
