% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamsApp.R
\name{streamsInput}
\alias{streamsInput}
\alias{streamsServer}
\alias{streamsApp}
\title{UI Component for StreamStats Leaflet Explorer}
\usage{
streamsInput(id)

streamsServer(
  id,
  default_lat = 43.0731,
  default_lng = -89.4012,
  default_zoom = 9
)

streamsApp()
}
\arguments{
\item{id}{Shiny module ID}

\item{default_lat}{Initial latitude for the map view (default: 43.0731)}

\item{default_lng}{Initial longitude for the map view (default: -89.4012)}

\item{default_zoom}{Initial map zoom level (default: 9)}
}
\description{
UI Component for StreamStats Leaflet Explorer

Server Logic for StreamStats Leaflet Explorer

Run the Standalone StreamStats Explorer App
}
