% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.r
\name{topologies}
\alias{topologies}
\alias{pixel_topology}
\alias{reach_topology}
\title{Build topologies
Build pixel and reach topologies for delineated streams}
\usage{
pixel_topology(x, drainage, stream)

reach_topology(x, Tp, stream)
}
\arguments{
\item{x}{A \link[raster:stack]{raster::stack}, such as one created by \code{\link[=delineate]{delineate()}}, or specify layers separately with \code{drain} and \code{stream}.}

\item{drainage}{Optional, ignored if \code{x} is provided; a drainage direction raster}

\item{stream}{Optional, ignored if \code{x} is provided; a delineated stream raster, all non-stream cells must be \code{NA}}

\item{Tp}{Topology for pixels in the network, e.g., the output from \code{\link[=pixel_topology]{pixel_topology()}}.}
}
\value{
A \link[Matrix:sparseMatrix]{Matrix::sparseMatrix} giving the pixel or reach topology
}
\description{
Build topologies
Build pixel and reach topologies for delineated streams
}
\details{
The topology is a square matrix showing immediate adjacency between pixels/reaches. Each row/column
in the topology is a single pixel/reach. Zero entries indicate no adjacency between two nodes. A non-zero entry
for row \code{i} column \code{j} indicates that \code{i} is upstream of \code{j}. The value in the entry is the length of the
upstream pixel/reach.
}
\examples{
\donttest{
    data(kamp_dem)
    Tp = pixel_topology(kamp_dem)
    Tr = reach_topology(kamp_dem, Tp)
}
}
