% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.r
\docType{package}
\name{watershed_package}
\alias{watershed_package}
\title{watershed: Tools for Watershed Delineation}
\description{
This is a simplified set of tools for deriving a river network from digital elevation data.
The package relies on GRASS GIS (v. 7.4 or 7.6) to do the heavy lifting. An installation of
GRASS and the rgrass7 package is required for watershed to function. Note that this is meant
to be a simplified workflow for watershed delineation using only R code; for more options
users can use rgrass7 to access GRASS GIS functions directly.
}
\section{Key functions}{

\itemize{
\item \code{\link[=foo]{foo()}} A function for doing foo
}
}

\section{Datasets}{

\itemize{
\item \link{ybbs_dem} An example digital elevation model for the Ybbs catchment in Austria
}
}

